requirements.txt
asgiref==3.10.0
Django==5.2.8
django-environ==0.12.0
django-ratelimit==4.1.0
django-ratelimit2==0.4.2
sqlparse==0.5.3
